以下の不具合を修正した。
・平均処理時間をlongからdoubleに変更した。
・最大リトライ回数を指定していない場合に、例外発生時にhandleRetryOver()が呼び出されない不具合を修正した。

+++ src/main/java/jp/ossc/nimbus/service/queue/DistributedQueueHandlerContainerServiceMBean.java
@@ -248,7 +248,7 @@ public interface DistributedQueueHandlerContainerServiceMBean extends ServiceBas
      *
      * @return 平均処理時間[ms]
      */
-    public long getAverageHandleProcessTime();
+    public double getAverageHandleProcessTime();
     
     /**
      * 再開する。<p>
